Convert Python to Families
