+++ python/orchestrate/_internal/version.py
@@ -1,0 +1,3 @@
+from importlib import metadata
+
+__version__ = metadata.version("orchestrate-api")
