+++ python/orchestrate/insight.py
@@ -1,29 +1,11 @@
-from typing import Literal, TypedDict, Union
-from orchestrate._internal.fhir import (
-    Measure,
-    MeasureReport,
-    OperationOutcome,
-    Patient,
-    RiskAssessment,
-    ValueSet,
+from orchestrate._internal.insight import (
+    InsightBundle,
+    InsightBundleEntry,
+    InsightRiskProfileResponse,
 )
 
-
-class InsightBundleEntry(TypedDict):
-    resource: Union[
-        Patient,
-        MeasureReport,
-        Measure,
-        ValueSet,
-        RiskAssessment,
-        OperationOutcome,
-    ]
-
-
-class InsightBundle(TypedDict):
-    id: str
-    resourceType: Literal["Bundle"]
-    entry: list[InsightBundleEntry]
-
-
-InsightRiskProfileResponse = InsightBundle
+__all__ = [
+    "InsightBundle",
+    "InsightBundleEntry",
+    "InsightRiskProfileResponse",
+]
